Set Position structure fields to int instead of size_t

+++ src/Utilities/to2d.cpp
@@ -11,8 +11,8 @@
 
 ind::Position ind::to2d(const irr::core::vector3df& pos3d)
 {
-    size_t x = (int)((pos3d.X + TILE_SIZE / 2) / TILE_SIZE);
-    size_t y = (int)((pos3d.Z + TILE_SIZE / 2) / TILE_SIZE);
+    int x = (int)((pos3d.X + TILE_SIZE / 2) / TILE_SIZE);
+    int y = (int)((pos3d.Z + TILE_SIZE / 2) / TILE_SIZE);
 
     return {x, y};
 }
